(split)http://bugzilla.liquidsys.com/show_bug.cgi?id=3693 bug: 3693
Copied from Perforce
 Change: 11611

+++ src/java/com/zimbra/cs/service/FileUploadServlet.java
@@ -128,7 +128,7 @@ public class FileUploadServlet extends ZimbraServlet {
         Upload up = null;
         synchronized (mPending) {
             up = (Upload) mPending.remove(uploadId);
-            if (!accountId.equals(up.accountId)) {
+            if (up != null && !accountId.equals(up.accountId)) {
                 mLog.warn("deleteUploads(" + accountId + ", " + uploadId +
                     ").  Found mismatched accountId: " + up);
                 mPending.put(uploadId, up);
